+++ ChangeLog.md
@@ -4,6 +4,8 @@
 
 * Better matching of constraints generated for GADTs across GHC versions
 * Added `dataDCompat`
+* Added `resolveInfixT` which uses reified fixity information to resolve `UInfixT`
+  constructors.
 
 ## 0.1.0.0  -- 2017-04-26